% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\docType{data}
\name{chipseq}
\alias{chipseq}
\title{chipseq}
\format{An object of class \code{data.frame} with 1000 rows and 2 columns.}
\source{
"https://www.encodeproject.org/"
}
\usage{
chipseq
}
\description{
A data frame of two columns. Each column represents a ChIP-seq data
replicates. The values are the read counts at different genomic locations.
}
\references{
To be added.
}
\keyword{datasets}
