% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.clayton.R
\name{d.clayton}
\alias{d.clayton}
\title{function to caluculate the density of Clayton copula}
\usage{
d.clayton(u, v, beta)
}
\arguments{
\item{u}{The cumulative density of a random variable. It can be calculated using `empdist` function.}

\item{v}{The cumulative density of another random variable.}

\item{beta}{The shape parameter for Clayton copula.}
}
\value{
The density of Clayton Copula.
}
\description{
function to caluculate the density of Clayton copula
}
