% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.gaussian.R
\name{d.gaussian}
\alias{d.gaussian}
\title{function to caluculate the density of Guassian copula}
\usage{
d.gaussian(u, v, rho)
}
\arguments{
\item{u}{The cumulative density of a random variable. It can be calculated using `empdist` function.}

\item{v}{The cumulative density of another random variable.}

\item{rho}{The correlation paramter of Guassian copula.}
}
\value{
The density of Clayton Copula.
}
\description{
function to caluculate the density of Guassian copula
}
