% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empdist.R
\name{empdist}
\alias{empdist}
\title{calculate the emperical cumulative density}
\usage{
empdist(x1, x2)
}
\arguments{
\item{x1}{A vector of signal values. For example, the read counts in genomic bins.}

\item{x2}{Another vector of same length signal values that to be test reproducibility against x1.}
}
\value{
A data frame with two columns and the same length with x1 and x2.
}
\description{
calculate the emperical cumulative density
}
