% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.TDR.R
\name{est.TDR}
\alias{est.TDR}
\title{EM algorithm to estimate the parameters of mixture model of survival Clayton
copula and Gaussian copula.}
\usage{
est.TDR(x1, x2, p = 0.25, rho = 0.2, beta = 2, max.it = 100,
  eps = 0.001)
}
\arguments{
\item{x1}{A vector of signal values. For example, the read counts in genomic bins.}

\item{x2}{Another vector of same length signal values that to be test reproducibility
against x1.}

\item{p}{The starting value of proportion of Guassion copula for the EM algorithm.
The default is 0.5.}

\item{rho}{The starting value of Guassian correlation parameter for the EM algorithm.
The default is 0.2. It can be any value betwen 0 and 1.}

\item{beta}{The starting value of Clayton shape parameter for the EM algorithm.
The default is 2. It can be any value greater than 0.}

\item{max.it}{The maximum number of iterations for EM algorithm. The default is 100.}

\item{eps}{The threhold of convergence. Iteration of EM stops if all the changes of
parameters are smaller than this threshold.}
}
\value{
A list of results including parameter traces,log likelihood trace and final
estimated parameters.
\itemize{
  \item{para.trace    }{The trace of parameters along the estimation iterations.}
   \item{likelihood.trace    }{The trace of log likelihood. The trace of log likelihood
   can use to monitor the estimation. It should be monotonically increasing afte
   first few iterations.}
   \item{para    }{The estimated parameters, including pi, rho, lambda.}
}
}
\description{
EM algorithm to estimate the parameters of mixture model of survival Clayton
copula and Gaussian copula.
}
\details{
This function estimates the parameters for the mixture copula of survival
Clayton and Gaussian.
}
\examples{
p <- 0.25
rho <- 0.2
beta <- 2
max.it <- 100
eps <- 0.001
data(chipseq)
TDR.out = est.TDR(chipseq$R1, chipseq$R2, p, rho, beta, max.it, eps)
TDR.out$para
}
\references{
To be added.
}
